perbaikan backend dan menyambungkan ke interface

+++ src/main/java/com/example/demo/Dokter/DokterModel.java
@@ -14,7 +14,7 @@ import lombok.experimental.FieldDefaults;
 @AllArgsConstructor
 @Data
 public class DokterModel {
-    Long idPegawai;
+    int idPegawai;
    
     String username;
     String password;
